feat(15-02): src/ confirmation path adapter + test rewrite for async Future
- Mark ReplyQueueLike in agent_run_service.py as deprecated (preserved for stream_service import)
- Add ConfirmationHookAdapter in stream_service.py bridging put_content/put_cancel to hook.resolve/cancel
- Rewrite test_confirmation.py for asyncio.Future model (13 tests: no-loop, with-loop, cross-thread, adapter)
- Update TestCrossPodReplyQueue in test_upstream_scenarios.py for async ConfirmationHook
- Fix timeout_sec type to float for sub-second test timeouts (cast to int for event emission)

+++ matmaster/hooks/confirmation.py
@@ -32,7 +32,7 @@ class ConfirmationHook(BaseHook):
         self,
         bus: MessageBus,
         *,
-        timeout_sec: int = 20,
+        timeout_sec: float = 20,
         confirm_tools: set[str] | None = None,
         source: str = "MatMaster",
     ) -> None:
@@ -64,7 +64,7 @@ class ConfirmationHook(BaseHook):
                 source=self._source,
                 question=f"Confirm tool call: {tool_call.name}?",
                 mode="timeout",
-                timeout_seconds=self._timeout_sec,
+                timeout_seconds=int(self._timeout_sec),
             )
         )
